move checking if TE is installed to fp8/init.py
