+++ src/nanotron/fp8/meta.py
@@ -1,17 +1,12 @@
-import warnings
 from dataclasses import dataclass
 from typing import Union
 
 import torch
+import transformer_engine as te  # noqa
+import transformer_engine_extensions as tex
 
 from nanotron.fp8.constants import DTYPE_TO_FP8_MAX
 from nanotron.fp8.tensor import convert_torch_dtype_to_te_dtype
-
-try:
-    import transformer_engine as te  # noqa
-    import transformer_engine_extensions as tex
-except ImportError:
-    warnings.warn("Please install Transformer engine for FP8 training!")
 
 
 @dataclass
