+++ tests/fp8/test_fp8_parameter.py
@@ -1,8 +1,6 @@
 import torch
-from nanotron.fp8.dtypes import DTypes
+from nanotron.fp8 import DTypes, FP8Parameter, FP8Tensor
 from nanotron.fp8.meta import FP8Meta
-from nanotron.fp8.parameter import FP8Parameter
-from nanotron.fp8.tensor import FP8Tensor
 
 
 def test_create_fp8_parameter():
